+++ libs/framework-core/source/ftrack_framework_core/event/__init__.py
@@ -46,7 +46,7 @@ class EventManager(object):
     '''Manages the events handling.'''
 
     def __repr__(self):
-        return '<EventManager:{}:{}>'.format(self.allow_remote_events, self.id)
+        return '<EventManager:{}:{}>'.format(self.mode, self.id)
 
     def __del__(self):
         self.logger.debug('Closing {}'.format(self))
@@ -71,18 +71,8 @@ class EventManager(object):
         return _connected
 
     @property
-    def allow_remote_events(self):
-        return self._allow_remote_events
-
-    @allow_remote_events.setter
-    def allow_remote_events(self, value):
-        if not isinstance(value, bool):
-            self.logger.error('allow_remote_events must be a boolean')
-            return
-        if value:
-            self._connect()
-            self._wait()
-        self._allow_remote_events = value
+    def mode(self):
+        return self._mode
 
     @property
     def publish(self):
@@ -123,13 +113,17 @@ class EventManager(object):
         self.session.event_hub.disconnect()
         self.session.close()
 
-    def __init__(self, session, allow_remote_events=True):
+    def __init__(self, session, mode=constants.event.LOCAL_EVENT_MODE):
         self.logger = logging.getLogger(
             __name__ + '.' + self.__class__.__name__
         )
         self._event_hub_thread = None
+        self._mode = mode
         self._session = session
-        self.allow_remote_events = allow_remote_events
+        if mode == constants.event.REMOTE_EVENT_MODE:
+            # TODO: Bring this back when API event hub properly can differentiate between local and remote mode
+            self._connect()
+            self._wait()
 
         # Initialize Publish and subscribe classes to be able to provide
         # predefined events.
@@ -137,13 +131,13 @@ class EventManager(object):
         self._subscribe_instance = Subscribe(self)
 
         self.logger.debug(
-            f'Initialising event manager {self} with remote events {self.allow_remote_events}'
+            f'Initialising event manager {self} with mode {self.mode}'
         )
 
-    def _publish(self, event, callback=None, remote=None):
+    def _publish(self, event, callback=None):
         '''Emit *event* and provide *callback* function, in local or remote *remote*.'''
 
-        if not remote:
+        if self.mode == constants.event.LOCAL_EVENT_MODE:
             result = self.session.event_hub.publish(
                 event,
                 synchronous=True,
@@ -162,13 +156,6 @@ class EventManager(object):
             return result
 
         else:
-            if not self.allow_remote_events:
-                self.logger.warning(
-                    'Remote events are not allowed. Please initialize the '
-                    'EventManager with allow_remote_events=True or set the '
-                    'allow_remote_Events property to true'
-                )
-                return
             self.session.event_hub.publish(event, on_reply=callback)
 
     def _subscribe(self, topic, callback):
@@ -197,7 +184,7 @@ class Publish(object):
         super(Publish, self).__init__()
         self._event_manager = event_manager
 
-    def _publish_event(self, event_topic, data, callback, remote=None):
+    def _publish_event(self, event_topic, data, callback):
         '''
         Common method that calls the private publish method from the
         event manager
@@ -206,11 +193,11 @@ class Publish(object):
             topic=event_topic, data=data
         )
         publish_result = self.event_manager._publish(
-            publish_event, callback=callback, remote=remote
+            publish_event, callback=callback
         )
         return publish_result
 
-    def discover_host(self, callback=None, remote=None):
+    def discover_host(self, callback=None):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.DISCOVER_HOST_TOPIC`
@@ -218,17 +205,10 @@ class Publish(object):
         # TODO: review this implementation, for now this one can't never be remote as it goes into a loop.
         data = None
         event_topic = constants.event.DISCOVER_HOST_TOPIC
-        return self._publish_event(
-            event_topic, data, callback, remote=False
-        )  # Change this once fixed.
+        return self._publish_event(event_topic, data, callback)
 
     def host_run_tool_config(
-        self,
-        host_id,
-        tool_config_reference,
-        client_options,
-        callback=None,
-        remote=None,
+        self, host_id, tool_config_reference, client_options, callback=None
     ):
         '''
         Publish an event with topic
@@ -240,7 +220,7 @@ class Publish(object):
             'client_options': client_options,
         }
         event_topic = constants.event.HOST_RUN_TOOL_CONFIG_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
     def host_run_ui_hook(
         self,
@@ -250,7 +230,6 @@ class Publish(object):
         client_options,
         payload,
         callback=None,
-        remote=None,
     ):
         '''
         Publish an event with topic
@@ -264,11 +243,9 @@ class Publish(object):
             'payload': payload,
         }
         event_topic = constants.event.HOST_RUN_UI_HOOK_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
-    def host_context_changed(
-        self, host_id, context_id, callback=None, remote=None
-    ):
+    def host_context_changed(self, host_id, context_id, callback=None):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.HOST_CONTEXT_CHANGED_TOPIC`
@@ -278,11 +255,9 @@ class Publish(object):
             'context_id': context_id,
         }
         event_topic = constants.event.HOST_CONTEXT_CHANGED_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
-    def client_context_changed(
-        self, host_id, context_id, callback=None, remote=None
-    ):
+    def client_context_changed(self, host_id, context_id, callback=None):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.CLIENT_CONTEXT_CHANGED_TOPIC`
@@ -292,11 +267,9 @@ class Publish(object):
             'context_id': context_id,
         }
         event_topic = constants.event.CLIENT_CONTEXT_CHANGED_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
-    def host_log_item_added(
-        self, host_id, log_item, callback=None, remote=None
-    ):
+    def host_log_item_added(self, host_id, log_item, callback=None):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.HOST_LOG_ITEM_ADDED_TOPIC`
@@ -307,15 +280,10 @@ class Publish(object):
         }
 
         event_topic = constants.event.HOST_LOG_ITEM_ADDED_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
     def host_run_ui_hook_result(
-        self,
-        host_id,
-        plugin_reference,
-        ui_hook_result,
-        callback=None,
-        remote=None,
+        self, host_id, plugin_reference, ui_hook_result, callback=None
     ):
         '''
         Publish an event with topic
@@ -328,11 +296,9 @@ class Publish(object):
         }
 
         event_topic = constants.event.HOST_UI_HOOK_RESULT_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
-    def client_signal_context_changed(
-        self, client_id, callback=None, remote=None
-    ):
+    def client_signal_context_changed(self, client_id, callback=None):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.CLIENT_SIGNAL_CONTEXT_CHANGED_TOPIC`
@@ -342,11 +308,9 @@ class Publish(object):
         }
 
         event_topic = constants.event.CLIENT_SIGNAL_CONTEXT_CHANGED_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
-    def client_signal_host_changed(
-        self, client_id, callback=None, remote=None
-    ):
+    def client_signal_host_changed(self, client_id, callback=None):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.CLIENT_SIGNAL_HOST_CHANGED_TOPIC`
@@ -356,11 +320,9 @@ class Publish(object):
         }
 
         event_topic = constants.event.CLIENT_SIGNAL_HOST_CHANGED_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
-    def client_notify_log_item_added(
-        self, client_id, log_item, callback=None, remote=None
-    ):
+    def client_notify_log_item_added(self, client_id, log_item, callback=None):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.CLIENT_NOTIFY_LOG_ITEM_ADDED_TOPIC`
@@ -371,15 +333,10 @@ class Publish(object):
         }
 
         event_topic = constants.event.CLIENT_NOTIFY_LOG_ITEM_ADDED_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
     def client_notify_ui_hook_result(
-        self,
-        client_id,
-        plugin_reference,
-        ui_hook_result,
-        callback=None,
-        remote=None,
+        self, client_id, plugin_reference, ui_hook_result, callback=None
     ):
         '''
         Publish an event with topic
@@ -392,11 +349,9 @@ class Publish(object):
         }
 
         event_topic = constants.event.CLIENT_NOTIFY_UI_HOOK_RESULT_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
-    def host_verify_plugins(
-        self, host_id, plugin_names, callback=None, remote=None
-    ):
+    def host_verify_plugins(self, host_id, plugin_names, callback=None):
         '''
         Publish an event with topic
         :const:`~ftrack_framework_core.constants.event.HOST_VERIFY_PLUGINS_TOPIC`
@@ -407,7 +362,7 @@ class Publish(object):
         }
 
         event_topic = constants.event.HOST_VERIFY_PLUGINS_TOPIC
-        return self._publish_event(event_topic, data, callback, remote=remote)
+        return self._publish_event(event_topic, data, callback)
 
 
 class Subscribe(object):