+++ projects/framework-maya/source/ftrack_framework_maya/__init__.py
@@ -116,7 +116,9 @@ def bootstrap_integration(framework_extensions_path):
     )
     # Create ftrack session and instantiate event manager
     session = ftrack_api.Session(auto_connect_event_hub=True)
-    event_manager = EventManager(session=session)
+    event_manager = EventManager(
+        session=session, mode=constants.event.LOCAL_EVENT_MODE
+    )
     logger.debug(f"framework_extensions_path:{framework_extensions_path}")
     # Instantiate registry
     registry_instance = Registry()